python_visual/requirements.txt
pandas==2.1.0
psycopg2-binary==2.9.9
matplotlib==3.8.0
